tests/deviceorientationabsolute_event-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>deviceorientationabsolute attributes</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Yunx Liu" href="mailto:yunx.liu@intel.com">
<link rel="help" href="http://w3c.github.io/deviceorientation/spec-source-orientation.html#deviceorientationabsolute">
<script src="./resources/testharness.js"></script>
<script src="./resources/testharnessreport.js"></script>
<p>Rotate the device to run the test.</p>
<div id="log"></div>
<script>
async_test(function(t) {
  window.addEventListener("deviceorientationabsolute", function(e) {
    t.step(function() {
      assert_true("alpha" in e);
      assert_true("beta" in e);
      assert_true("gamma" in e);
      assert_true("absolute" in e);
      assert_equals(typeof e.absolute, "boolean");
    });
    t.done();
  }, false);
}, document.title);
</script>
